requirements.txt
pyspark
boto3
pandas
numpy
ruff
python-dotenv
